Adding shlibs and lintian-overrides files
Due to the way libraries (i.e., plugins) are built by Midori (without
a proper SONAME), it is necessary (or at least recommended) to provide
a shlibs file that contains trustworthy information about the
libraries being installed on the system.

Unfortunately, it is also necessary to provide a lintian-overrides
file in order to silence some lintian warnings (mostly about the
shlibs mentioned above).  We'll have to work with upstream in order to
solve this problem.

Another lintian error that needed to be silenced was related to the
data/midori.swf file, present at the original tarball.  This file is
used just for testing (i.e., it is not installed in the system), and
is actually generated using the png2swf utility (from the swftools
package), which means it can be verified.