Better handling of self-signed certificates
Midori needs gcr/gnome-keyring in order to fully work with self-signed
TLS certificates.  This commit improves the error message that is
displayed when clicking on "Trust this website" does not work, and
also makes the Debian Midori package recommend installing
gnome-keyring.  This is the best we can do for now.

Closes: #799963