Improve security by making sure that we use a safe set of ciphers.
This comes from Debian bug #804196 (LP Bug #1517265).  When Midori
uses libwebkitgtk, it inherits the default value of the
G_TLS_GNUTLS_PRIORITY environment variable.  Unfortunately this value
is not safe enough because it still lets GnuTLS to enable to unsafe
ciphers.  Therefore, we need to set this variable ourselves and make
sure that Midori is using safe ciphers for TLS operations.

Closes: #804196